Hrm support: correctly calculate Background datapage 1 cumulative time. refactor to  small logging util to accept lambdas on behalf of slf4j

+++ src/main/java/be/glever/antplus/hrm/datapage/AbstractHRMDataPage.java
@@ -23,13 +23,13 @@ public abstract class AbstractHRMDataPage extends AbstractAntPlusDataPage {
      * Note that the ANT+ field size for this is 16 bit and resolution is 1/1024 seconds,
      * so the  value rolls over at 63.999 seconds ( ((1/1024)*(2^17)) - (1/1024)) ).
      */
-    public double getHeartBeatEventTime() {
+    public long getHeartBeatEventTime() {
         return calculateHeartBeatEventTime(getDataPageBytes()[4], getDataPageBytes()[5]);
     }
 
-    protected double calculateHeartBeatEventTime(byte bite1, byte bite2) {
+    protected long calculateHeartBeatEventTime(byte bite1, byte bite2) {
         double timeAnt = ByteUtils.fromUShort(bite1, bite2);
-        return (timeAnt / 1024) * 1000;
+        return (long) ((timeAnt / 1024) * 1000);
     }
 
     /**
